Put short IDE section near the top
This is probably a little controversial but here goes:
Every time I see or ask someone to download the seleniumIDE they go to the selenium downloads page, they read through the first couple of screens, they think ah, standalone server must be it and they try installing that.  After a few miss-steps like this some folks give up
My suggestion is to move the 'IDE" section up for better visibility.  However - to be very clear - I am NOT saying 'favor having the ide section first' because after all many people don't want the IDE they want the server.
My point is as follows:  The IDE paragraph and download is very short.  If it was placed first, novices would see it right away AND more advanced users would be more likely to: see it, know it doesn't apply to them, they see the selenium driver download right below it, a few lines down without scrolling.
So I am not saying that there are 'more' IDE users or that the IDE stuff should have priority or anything like that.
I am saying that the IDE part is currently buried and non-obvious which significantly effects novice user and this can addressed by moving the paragraph up and I think that doing so will not significantly impact users looking for the server download which will be visible on the same page without scrolling.