Adding new assets for the site to source control
This makes it easier for us to share them without needing a
Top Secret zip file being passed around everywhere. I've
not included the fonts, since the names are referenced in
the style guides and they can be downloaded when we're ready
to use them.

I've also only included the SVG versions of the logos and
nav assets. This should be fine, since all browsers support
SVG out of the box now.